Replace EERepeatTests with EERepeatActions
